# Improve performance and reduce memory use of mcx io