--- a/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py
+++ b/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py
@@ -3,7 +3,6 @@
 # SPDX-License-Identifier: MIT
 
 import numpy as np
-import struct
 import subprocess
 from simpa.utils import Tags, Settings
 from simpa.core.simulation_modules.optical_simulation_module import OpticalForwardModuleBase
@@ -78,8 +77,6 @@ class MCXAdapter(OpticalForwardModuleBase):
 
         # Read output
         results = self.read_mcx_output()
-
-        struct._clearcache()
 
         # clean temporary files
         self.remove_mcx_output()
@@ -217,27 +214,14 @@ class MCXAdapter(OpticalForwardModuleBase):
                                                                    'scattering_cm': scattering_cm,
                                                                    'anisotropy': anisotropy,
                                                                    'assumed_anisotropy': assumed_anisotropy})
-        op_array = np.asarray([absorption_mm, scattering_mm])
-
+        op_array = np.asarray([absorption_mm, scattering_mm], dtype=np.float32)
         [_, self.nx, self.ny, self.nz] = np.shape(op_array)
-
-        # create a binary of the volume
-
-        optical_properties_list = list(np.reshape(op_array, op_array.size, "F"))
-        del absorption_cm, absorption_mm, scattering_cm, scattering_mm, op_array
-        gc.collect()
-        mcx_input = struct.pack("f" * len(optical_properties_list), *optical_properties_list)
-        del optical_properties_list
-        gc.collect()
+        # # create a binary of the volume
         tmp_input_path = self.global_settings[Tags.SIMULATION_PATH] + "/" + \
                          self.global_settings[Tags.VOLUME_NAME] + ".bin"
         self.temporary_output_files.append(tmp_input_path)
-        with open(tmp_input_path, "wb") as input_file:
-            input_file.write(mcx_input)
-
-        del mcx_input, input_file
-        struct._clearcache()
-        gc.collect()
+        # numpy tofile writes in 'C' order, so writing the transpose gives Fortran order
+        op_array.T.tofile(tmp_input_path)
 
     def read_mcx_output(self, **kwargs) -> Dict:
         """
@@ -246,10 +230,8 @@ class MCXAdapter(OpticalForwardModuleBase):
         :param kwargs: dummy, used for class inheritance compatibility
         :return: `Dict` instance containing the MCX output
         """
-        with open(self.mcx_volumetric_data_file, 'rb') as f:
-            data = f.read()
-        data = struct.unpack('%df' % (len(data) / 4), data)
-        fluence = np.asarray(data).reshape([self.nx, self.ny, self.nz, self.frames], order='F')
+        shape = [self.nx, self.ny, self.nz, self.frames]
+        fluence = np.fromfile(self.mcx_volumetric_data_file, dtype=np.float32).reshape(shape, order='F')
         fluence *= 100  # Convert from J/mm^2 to J/cm^2
         if np.shape(fluence)[3] == 1:
             fluence = np.squeeze(fluence, 3)