- generate_mcx_bin_input():
  - write binary file of floats in 'F' format by writing transposed numpy array of floats in 'C' format using `tofile()`
- read_mcx_output():
  - read binary file of floats using `fromfile()`
- avoids intermediate creation of (slow & ram-heavy) python objects by struct.pack/unpack
- benchmark using `minimal_optical_simulation.py` (with `SPACING=0.1`):
  - `generate_mcx_bin_input()`:
    - runtime: 17.20s -> 5.75s
    - peak memory: 16GB -> 2.5GB
  - `read_mcx_output()`:
    - runtime: 4.12s -> 0.15s
    - peak memory: 4.9GB -> 0.4GB

Runtime before:

![main](https://github.com/IMSY-DKFZ/simpa/assets/12845624/d316660c-c333-4f4d-b6ff-97bf7d7329cb)

after:

![this_pr](https://github.com/IMSY-DKFZ/simpa/assets/12845624/0b47fed6-8f68-4315-be7a-ffb0ae843db3)

Memory use before:

```
Filename: /home/lkeegan/SSC/simpa/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py

Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
   202 3757.016 MiB 3757.016 MiB           1       @profile
   203                                             def generate_mcx_bin_input(self,
   204                                                                        absorption_cm: np.ndarray,
   205                                                                        scattering_cm: np.ndarray,
   206                                                                        anisotropy: np.ndarray,
   207                                                                        assumed_anisotropy: np.ndarray) -> None:
   208                                                 """
   209                                                 generates binary file containing volume scattering and absorption as input for MCX
   210                                         
   211                                                 :param absorption_cm: Absorption in units of per centimeter
   212                                                 :param scattering_cm: Scattering in units of per centimeter
   213                                                 :param anisotropy: Dimensionless scattering anisotropy
   214                                                 :param assumed_anisotropy:
   215                                                 :return: None
   216                                                 """
   217 5404.977 MiB 1647.961 MiB           2           absorption_mm, scattering_mm = self.pre_process_volumes(**{'absorption_cm': absorption_cm,
   218 3757.016 MiB    0.000 MiB           1                                                                      'scattering_cm': scattering_cm,
   219 3757.016 MiB    0.000 MiB           1                                                                      'anisotropy': anisotropy,
   220 3757.016 MiB    0.000 MiB           1                                                                      'assumed_anisotropy': assumed_anisotropy})
   221 7052.914 MiB 1647.938 MiB           1           op_array = np.asarray([absorption_mm, scattering_mm])
   222                                         
   223 7052.914 MiB    0.000 MiB           1           [_, self.nx, self.ny, self.nz] = np.shape(op_array)
   224                                         
   225                                                 # create a binary of the volume
   226                                         
   227 15246.820 MiB 8193.906 MiB           1           optical_properties_list = list(np.reshape(op_array, op_array.size, "F"))
   228 11950.961 MiB -3295.859 MiB           1           del absorption_cm, absorption_mm, scattering_cm, scattering_mm, op_array
   229 11950.961 MiB    0.000 MiB           1           gc.collect()
   230 19775.926 MiB 7824.965 MiB           1           mcx_input = struct.pack("f" * len(optical_properties_list), *optical_properties_list)
   231 11511.004 MiB -8264.922 MiB           1           del optical_properties_list
   232 11511.004 MiB    0.000 MiB           1           gc.collect()
   233 11511.004 MiB    0.000 MiB           3           tmp_input_path = self.global_settings[Tags.SIMULATION_PATH] + "/" + \
   234 11511.004 MiB    0.000 MiB           2                            self.global_settings[Tags.VOLUME_NAME] + ".bin"
   235 11511.004 MiB    0.000 MiB           1           self.temporary_output_files.append(tmp_input_path)
   236 11511.004 MiB    0.000 MiB           2           with open(tmp_input_path, "wb") as input_file:
   237 11511.004 MiB    0.000 MiB           1               input_file.write(mcx_input)
   238                                         
   239 10687.027 MiB -823.977 MiB           1           del mcx_input, input_file
   240 3683.234 MiB -7003.793 MiB           1           struct._clearcache()
   241 3683.234 MiB    0.000 MiB           1           gc.collect()
```

```
Filename: /home/lkeegan/SSC/simpa/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py

Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
   243 3683.234 MiB 3683.234 MiB           1       @profile
   244                                             def read_mcx_output(self, **kwargs) -> Dict:
   245                                                 """
   246                                                 reads the temporary output generated with MCX
   247                                         
   248                                                 :param kwargs: dummy, used for class inheritance compatibility
   249                                                 :return: `Dict` instance containing the MCX output
   250                                                 """
   251 4095.484 MiB    0.000 MiB           2           with open(self.mcx_volumetric_data_file, 'rb') as f:
   252 4095.484 MiB  412.250 MiB           1               data = f.read()
   253 7815.680 MiB 3720.195 MiB           1           data = struct.unpack('%df' % (len(data) / 4), data)
   254 8639.422 MiB  823.742 MiB           1           fluence = np.asarray(data).reshape([self.nx, self.ny, self.nz, self.frames], order='F')
   255 8639.422 MiB    0.000 MiB           1           fluence *= 100  # Convert from J/mm^2 to J/cm^2
   256 8639.422 MiB    0.000 MiB           1           if np.shape(fluence)[3] == 1:
   257 8639.422 MiB    0.000 MiB           1               fluence = np.squeeze(fluence, 3)
   258 8639.422 MiB    0.000 MiB           1           results = dict()
   259 8639.422 MiB    0.000 MiB           1           results[Tags.DATA_FIELD_FLUENCE] = fluence
   260 8639.422 MiB    0.000 MiB           1           return results
```
after:
```
Filename: /home/lkeegan/SSC/simpa/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py

Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
   202 3756.855 MiB 3756.855 MiB           1       @profile
   203                                             def generate_mcx_bin_input(self,
   204                                                                        absorption_cm: np.ndarray,
   205                                                                        scattering_cm: np.ndarray,
   206                                                                        anisotropy: np.ndarray,
   207                                                                        assumed_anisotropy: np.ndarray) -> None:
   208                                                 """
   209                                                 generates binary file containing volume scattering and absorption as input for MCX
   210                                         
   211                                                 :param absorption_cm: Absorption in units of per centimeter
   212                                                 :param scattering_cm: Scattering in units of per centimeter
   213                                                 :param anisotropy: Dimensionless scattering anisotropy
   214                                                 :param assumed_anisotropy:
   215                                                 :return: None
   216                                                 """
   217 5404.812 MiB 1647.957 MiB           2           absorption_mm, scattering_mm = self.pre_process_volumes(**{'absorption_cm': absorption_cm,
   218 3756.855 MiB    0.000 MiB           1                                                                      'scattering_cm': scattering_cm,
   219 3756.855 MiB    0.000 MiB           1                                                                      'anisotropy': anisotropy,
   220 3756.855 MiB    0.000 MiB           1                                                                      'assumed_anisotropy': assumed_anisotropy})
   221 6228.555 MiB  823.742 MiB           1           op_array = np.asarray([absorption_mm, scattering_mm], dtype=np.float32)
   222 6228.555 MiB    0.000 MiB           1           [_, self.nx, self.ny, self.nz] = np.shape(op_array)
   223                                                 # # create a binary of the volume
   224 6228.555 MiB    0.000 MiB           3           tmp_input_path = self.global_settings[Tags.SIMULATION_PATH] + "/" + \
   225 6228.555 MiB    0.000 MiB           2                            self.global_settings[Tags.VOLUME_NAME] + ".bin"
   226 6228.555 MiB    0.000 MiB           1           self.temporary_output_files.append(tmp_input_path)
   227                                                 # numpy tofile writes in 'C' order, so writing the transpose gives Fortran order
   228 6228.555 MiB    0.000 MiB           1           op_array.T.tofile(tmp_input_path)
```

```
Filename: /home/lkeegan/SSC/simpa/simpa/core/simulation_modules/optical_simulation_module/optical_forward_model_mcx_adapter.py

Line #    Mem usage    Increment  Occurrences   Line Contents
=============================================================
   230 3756.859 MiB 3756.859 MiB           1       @profile
   231                                             def read_mcx_output(self, **kwargs) -> Dict:
   232                                                 """
   233                                                 reads the temporary output generated with MCX
   234                                         
   235                                                 :param kwargs: dummy, used for class inheritance compatibility
   236                                                 :return: `Dict` instance containing the MCX output
   237                                                 """
   238 3756.859 MiB    0.000 MiB           1           shape = [self.nx, self.ny, self.nz, self.frames]
   239 4168.633 MiB  411.773 MiB           1           fluence = np.fromfile(self.mcx_volumetric_data_file, dtype=np.float32).reshape(shape, order='F')
   240 4168.633 MiB    0.000 MiB           1           fluence *= 100  # Convert from J/mm^2 to J/cm^2
   241 4168.633 MiB    0.000 MiB           1           if np.shape(fluence)[3] == 1:
   242 4168.633 MiB    0.000 MiB           1               fluence = np.squeeze(fluence, 3)
   243 4168.633 MiB    0.000 MiB           1           results = dict()
   244 4168.633 MiB    0.000 MiB           1           results[Tags.DATA_FIELD_FLUENCE] = fluence
   245 4168.633 MiB    0.000 MiB           1           return results
```